+++ eslint.config.mjs
@@ -6,4 +6,9 @@ export default defineConfig(
     eslint.configs.recommended,
     tseslint.configs.strict,
     tseslint.configs.stylistic,
+    { // https://github.com/eslint/eslint/issues/20272
+        rules: {
+            '@typescript-eslint/unified-signatures': 'off',
+        },
+    },
 );