+++ packages/graphico/CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG
 
+## 0.0.3
+
+- Add animation example
+- Expose `Canvas.width` and `Canvas.height`
+- Change `setInterval` to `requestAnimationFrame`
+- Create and expose several new functions in the `Canvas` class:
+    - `getMousePosition`
+    - `getPixel`
+    - `setPixel`
+    - `screenshot`
+- Main animation `loop()` function can return an array of `Drawable` to automatically render
+
 ## 0.0.2
 
 - Add first iteration of `Canvas` class